Moved validateTokenSecurity to a different file
